requirements.txt
Flask>=3.0,<4.0
flask-restx>=1.3,<2.0
Flask-Migrate>=4.0,<5.0
Flask-SQLAlchemy>=3.1,<4.0
waitress>=3.0,<4.0
SQLAlchemy>=2.0,<3.0
dynaconf>=3.2,<4.0
APScheduler>=3.10,<4.0
requests>=2.31,<3.0
PyYAML>=6.0
signalrcore>=0.9,<1.0
dogpile.cache>=1.3,<2.0
python-dateutil>=2.8
tzlocal>=5.0
pretty-errors>=1.2
mutagen>=1.47
